+++ packages/slice-machine/src/features/customTypes/customTypesBuilder/CustomTypeProvider.tsx
@@ -11,7 +11,6 @@ import {
 
 import { CustomTypeUpdateMeta, getState, updateCustomType } from "@/apiClient";
 import { getFormat } from "@/domain/customType";
-import { revalidateGetCustomTypes } from "@/features/customTypes/customTypesTable/useCustomTypes";
 import { useAutoSync } from "@/features/sync/AutoSyncProvider";
 import { ActionQueueStatus, useActionQueue } from "@/hooks/useActionQueue";
 import useSliceMachineActions from "@/modules/useSliceMachineActions";
@@ -67,13 +66,12 @@ export function CustomTypeProvider(props: CustomTypeProviderProps) {
         // Refresh the store with the latest server state to get the updated
         // custom types
         stableRefreshState(await getState());
-        revalidateGetCustomTypes(format);
 
         syncChanges();
         onSaveCallback?.();
       });
     },
-    [setNextAction, stableRefreshState, syncChanges, format],
+    [setNextAction, stableRefreshState, syncChanges],
   );
 
   const contextValue: CustomTypeContext = useMemo(
