+++ packages/slice-machine/src/features/customTypes/customTypesTable/useCustomTypesAutoRevalidation.tsx
@@ -1,56 +1,12 @@
-import {
-  revalidateData,
-  updateData,
-  useRequest,
-} from "@prismicio/editor-support/Suspense";
 import type { CustomType } from "@prismicio/types-internal/lib/customtypes";
 import type { CustomTypeFormat } from "@slicemachine/manager";
-import { useCallback, useEffect } from "react";
+import { useEffect } from "react";
 import { useSelector } from "react-redux";
 
 import { CustomTypes } from "@/legacy/lib/models/common/CustomType";
 import { hasLocal } from "@/legacy/lib/models/common/ModelData";
-import { managerClient } from "@/managerClient";
 import { selectAllCustomTypes } from "@/modules/availableCustomTypes";
 import type { SliceMachineStoreType } from "@/redux/type";
-
-type UseCustomTypesReturnType = {
-  customTypes: CustomType[];
-  updateCustomTypes: (customTypes: CustomType[]) => void;
-};
-
-export function useCustomTypes(
-  format?: CustomTypeFormat,
-): UseCustomTypesReturnType {
-  const updateCustomTypes = useCallback(
-    (data: CustomType[]) => updateData(getCustomTypes, [format], data),
-    [format],
-  );
-
-  return {
-    customTypes: useRequest(getCustomTypes, [format]),
-    updateCustomTypes,
-  };
-}
-
-export async function getCustomTypes(
-  format?: CustomTypeFormat,
-): Promise<CustomType[]> {
-  const { errors, models } = await managerClient.customTypes.readAllCustomTypes(
-    format ? { format } : undefined,
-  );
-
-  if (errors.length > 0) {
-    throw errors;
-  }
-
-  return models.map(({ model }) => model);
-}
-
-export function revalidateGetCustomTypes(format?: CustomTypeFormat) {
-  void revalidateData(getCustomTypes, []);
-  void revalidateData(getCustomTypes, [format]);
-}
 
 /**
  * TODO: DT-1363 - Update the way to have new data without Redux by revalidating
